// src/types/Resultado.tsx
export interface Resultado {
  mesesTrabalhados: number;

  motivoRescisao: string;

  saldoSalario: number;

  valorFeriasProporcionais: number;

  umTercoFerias: number;

  feriasVencidasTotal: number;

  decimoTerceiro: number;

  avisoIndenizado: number;

  saldoFgts: number;

  fgtsSobreSaldo: number;

  fgtsSobre13: number;

  fgtsSobreAviso: number;

  multaFgts: number;

  totalRecisao: number;

  decimoTerceiroSobreAviso: number;

  feriasProporcionaisSobreAviso: number;

  umTercoFeriasSobreAviso: number;
}